refactor static content loading
