+++ services/core-web/src/components/dashboard/incidentsHomePage/IncidentsHomePage.js
@@ -5,12 +5,6 @@ import { destroy } from "redux-form";
 import { debounce, isEmpty } from "lodash";
 import queryString from "query-string";
 import PropTypes from "prop-types";
-import {
-  fetchRegionOptions,
-  fetchMineComplianceCodes,
-  fetchMineIncidentStatusCodeOptions,
-  fetchMineIncidentCategoryCodeOptions,
-} from "@/actionCreators/staticContentActionCreator";
 import { modalConfig } from "@/components/modalContent/config";
 import { openModal, closeModal } from "@/actions/modalActions";
 import {
@@ -32,7 +26,6 @@ import { fetchIncidents, updateMineIncident } from "@/actionCreators/incidentAct
 import { getIncidents, getIncidentPageData } from "@/selectors/incidentSelectors";
 import { IncidentsTable } from "./IncidentsTable";
 import * as router from "@/constants/routes";
-import { fetchInspectors } from "@/actionCreators/partiesActionCreator";
 import IncidentsSearch from "./IncidentsSearch";
 import { formatParamStringToArray } from "@/utils/helpers";
 import * as ModalContent from "@/constants/modalContent";
@@ -45,19 +38,14 @@ import * as Strings from "@/constants/strings";
  */
 
 const propTypes = {
-  fetchMineComplianceCodes: PropTypes.func.isRequired,
-  fetchRegionOptions: PropTypes.func.isRequired,
   openModal: PropTypes.func.isRequired,
   closeModal: PropTypes.func.isRequired,
   destroy: PropTypes.func.isRequired,
-  fetchInspectors: PropTypes.func.isRequired,
   fetchIncidents: PropTypes.func.isRequired,
   inspectors: CustomPropTypes.groupOptions.isRequired,
   incidentPageData: CustomPropTypes.incidentPageData.isRequired,
   incidents: PropTypes.arrayOf(CustomPropTypes.incident).isRequired,
   updateMineIncident: PropTypes.func.isRequired,
-  fetchMineIncidentStatusCodeOptions: PropTypes.func.isRequired,
-  fetchMineIncidentCategoryCodeOptions: PropTypes.func.isRequired,
   location: PropTypes.shape({ search: PropTypes.string }).isRequired,
   mineRegionOptions: CustomPropTypes.options.isRequired,
   history: PropTypes.shape({ push: PropTypes.func }).isRequired,
@@ -140,11 +128,6 @@ export class IncidentsHomePage extends Component {
     this.props.fetchIncidents(this.state.params).then(() => {
       this.setState({ incidentsLoaded: true });
     });
-    this.props.fetchInspectors();
-    this.props.fetchMineComplianceCodes();
-    this.props.fetchRegionOptions();
-    this.props.fetchMineIncidentStatusCodeOptions();
-    this.props.fetchMineIncidentCategoryCodeOptions();
   }
 
   componentWillReceiveProps(nextProps) {
@@ -372,11 +355,6 @@ const mapDispatchToProps = (dispatch) =>
       destroy,
       openModal,
       closeModal,
-      fetchRegionOptions,
-      fetchMineComplianceCodes,
-      fetchMineIncidentStatusCodeOptions,
-      fetchMineIncidentCategoryCodeOptions,
-      fetchInspectors,
     },
     dispatch
   );
