+++ services/core-web/src/components/parties/RelationshipProfile.js
@@ -6,10 +6,7 @@ import PropTypes from "prop-types";
 import { Link } from "react-router-dom";
 import { Tabs, Table } from "antd";
 import { isEmpty } from "lodash";
-import {
-  fetchPartyRelationshipTypes,
-  fetchPartyRelationships,
-} from "@/actionCreators/partiesActionCreator";
+import { fetchPartyRelationships } from "@/actionCreators/partiesActionCreator";
 import { fetchMineRecordById } from "@/actionCreators/mineActionCreator";
 import { getPartyRelationshipTypesList, getPartyRelationships } from "@/selectors/partiesSelectors";
 import { getMines } from "@/selectors/mineSelectors";
@@ -28,7 +25,6 @@ const { TabPane } = Tabs;
 
 const propTypes = {
   fetchMineRecordById: PropTypes.func.isRequired,
-  fetchPartyRelationshipTypes: PropTypes.func.isRequired,
   fetchPartyRelationships: PropTypes.func.isRequired,
   partyRelationships: PropTypes.arrayOf(CustomPropTypes.partyRelationship),
   partyRelationshipTypes: PropTypes.arrayOf(CustomPropTypes.dropdownListItem),
@@ -74,9 +70,6 @@ export class RelationshipProfile extends Component {
     const mine = this.props.mines[id];
 
     // Fetch any props not provided
-    if (this.props.partyRelationshipTypes.length === 0) {
-      this.props.fetchPartyRelationshipTypes();
-    }
     if (this.props.partyRelationships.length === 0) {
       this.props.fetchPartyRelationships({
         mine_guid: id,
@@ -246,7 +239,6 @@ const mapDispatchToProps = (dispatch) =>
   bindActionCreators(
     {
       fetchMineRecordById,
-      fetchPartyRelationshipTypes,
       fetchPartyRelationships,
     },
     dispatch
